services/customer_web/requirements.txt
flask>=3.0
flask-socketio>=5.3
python-socketio>=5.11
eventlet>=0.35
requests>=2.31
pyyaml>=6.0
